Refactor viewer into separate fileviewer and treeviewer modules
Split the monolithic viewer.rs into three focused modules:
- fileviewer.rs: FileViewer struct and text content navigation
- treeviewer.rs: TreeViewer struct and AST tree navigation
- viewer.rs: Viewer trait and ViewerEvent enum

This improves code organization and maintainability by separating
concerns and making each component's responsibilities clearer.
